src/entities/Counter/ui/Counter.tsx
import React from 'react';
import { useTranslation } from 'react-i18next';
import { useDispatch, useSelector } from 'react-redux';
import { Button } from 'shared/ui/Button/Button';
import { getCounterValue } from '../model/selectors/getCounterValue/getCounterValue';
import { counterActions } from '../model/slice/counterSlice';

export const Counter: React.FC = () => {
  const dispatch = useDispatch();
  const counterValue = useSelector(getCounterValue);
  const { t } = useTranslation();

  const increment = () => {
    dispatch(counterActions.increment());
  };

  const decrement = () => {
    dispatch(counterActions.decrement());
  };

  return (
    <div>
      <h1 data-testid="counter-value-title">
        {counterValue}
      </h1>
      <Button data-testid="counter-increment-btn" onClick={increment}>{t('increment')}</Button>
      <Button data-testid="counter-decrement-btn" onClick={decrement}>{t('decrement')}</Button>
    </div>
  );
};
